Fix positive percentage type to decimal.

+++ lib/covid_scraper/entities/date_positive_test.rb
@@ -4,7 +4,7 @@ module CovidScraper
   module Entities
     class DatePositiveTest < ROM::Struct
       attribute :country, Types::Coercible::String
-      attribute :positive_percentage, Types::Coercible::Integer
+      attribute :positive_percentage, Types::Coercible::Decimal
       attribute :date, Types::Nominal::Date
     end
   end
